requirements.txt
adb-shell==0.4.4
aioquic-mitmproxy==0.9.21.1
annotated-types==0.7.0
ansible==7.7.0
ansible-core==2.14.11
argcomplete==2.0.0
asgiref==3.7.2
asttokens==2.4.1
attrs==23.1.0
Beaker==1.12.1
beautifulsoup4==4.12.3
blessed==1.20.0
blinker==1.6.2
blivet==3.7.1
blivet-gui==2.5.0
Brlapi==0.8.4
Brotli==1.1.0
bs4==0.0.1
cachelib==0.13.0
cairocffi==1.3.0
CairoSVG==2.7.0
certifi==2023.5.7
cffi==1.15.1
chardet==5.2.0
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
comm==0.2.1
crypto==1.4.1
cryptography==41.0.5
cssselect==1.2.0
cssselect2==0.7.0
cupshelpers==1.0
dasbus==1.7
DateTime==5.1
dbus-python==1.3.2
debugpy==1.8.1
decorator==4.4.2
defusedxml==0.7.1
distro==1.8.0
dnf==4.19.2
EasyProcess==1.1
ensta==5.2.9
entrypoint2==1.1
exceptiongroup==1.1.3
executing==2.0.1
fake-useragent==1.2.1
Faker==25.0.1
fedora-third-party==0.10
ffmpeg-python==0.2.0
file-magic==0.4.0
Flask==2.3.3
Flask-Session==0.8.0
fros==1.1
future==1.0.0
h11==0.14.0
h2==4.1.0
hpack==4.0.0
html5lib==1.1
humanize==3.13.1
hyperframe==6.0.1
idna==3.7
imageio==2.34.1
imageio-ffmpeg==0.4.9
initial-setup==0.3.97
inquirer==3.1.3
instagrapi==2.1.2
ipykernel==6.29.3
ipython==8.22.2
itsdangerous==2.1.2
jedi==0.19.1
jeepney==0.8.0
Jinja2==3.1.2
jupyter_client==8.6.0
jupyter_core==5.7.1
kaitaistruct==0.10
langtable==0.0.65
ldap3==2.9.1
libcomps==0.1.20
libdnf==0.73.1
loguru==0.7.2
lxml==4.9.2
Mako==1.2.3
markdown-it-py==3.0.0
MarkupSafe==2.1.2
matplotlib-inline==0.1.6
mdurl==0.1.2
mechanize==0.4.8
memory-profiler==0.61.0
MIDIUtil==1.2.1
mitmproxy==10.1.3
mitmproxy_rs==0.4.1
MouseInfo==0.1.3
moviepy==1.0.3
msgpack==1.0.7
msgspec==0.18.6
mss==9.0.1
Naked==0.1.32
nest-asyncio==1.6.0
nftables==0.1
ntplib==0.4.0
numpy==1.26.2
olefile==0.46
opencv-python==4.8.1.78
outcome==1.2.0
packaging==23.0
parso==0.8.3
passlib==1.7.4
Paste==3.5.2
pexpect==4.8.0
pid==2.2.3
Pillow==10.1.0
platformdirs==4.2.0
ply==3.11
productmd==1.38
proglog==0.1.10
prompt-toolkit==3.0.43
protobuf==4.25.0
psutil==5.9.5
ptyprocess==0.7.0
publicsuffix2==2.20191221
pure-eval==0.2.2
pwquality==1.4.5
pyasn1==0.5.0
pyasn1-modules==0.3.0
PyAutoGUI==0.9.54
pycairo==1.23.0
pycodestyle==2.5.0
pycparser==2.20
pycrypto==2.6.1
pycryptodome==3.18.0
pycryptodomex==3.20.0
pycups==2.0.4
pycurl==7.45.2
pydantic==2.7.1
pydantic_core==2.18.2
pydub==0.25.1
pyenchant==3.2.2
pyfakefs==5.3.0
pyfiglet==0.8.post1
PyGetWindow==0.0.9
Pygments==2.15.1
PyGObject==3.44.2
pykickstart==3.47
pylsqpack==0.3.18
pymailtm==1.1.1
PyMsgBox==1.0.9
pyngrok==6.1.0
pyOpenSSL==23.3.0
pyotp==2.9.0
pyparsing==3.1.1
pyparted==3.12.0
pyperclip==1.8.2
pyquery==2.0.0
PyRect==0.2.0
pyscreenshot==3.1
PyScreeze==0.1.29
PySocks==1.7.1
pystray==0.17.3
pystyle==2.9
python-augeas==1.1.0
python-dateutil==2.8.2
python-dotenv==1.0.1
python-editor==1.0.4
python-gettext==4.0
python-manatools==0.0.4
python-meh==0.51
python-xlib==0.33
python3-xlib==0.15
pyttsx3==2.90
pytube==15.0.0
pytweening==1.0.7
pytz==2023.3
pyudev==0.24.0
PyVirtualDisplay==3.0
pywhatkit==5.4
pyxdg==0.27
PyYAML==6.0
pyzmq==25.1.2
random-strings==0.1.1
random-username==1.0.2
readchar==4.0.5
regex==2024.4.16
requests==2.31.0
requests-file==1.5.1
requests-ftp==0.3.1
resolvelib==0.5.5
rich==13.4.2
rpm==4.18.2
rsa==4.9
ruamel.yaml==0.18.5
ruamel.yaml.clib==0.2.8
selenium==4.23.1
selinux==3.5
sepolicy==3.5
service-identity==23.1.0
setools==4.4.3
setroubleshoot @ file:///builddir/build/BUILD/setroubleshoot-3.3.33/src
shellescape==3.8.1
simpleaudio==1.0.4
simpleline==1.9.0
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
sos==4.6.0
soundfile==0.12.1
soupsieve==2.4.1
speedtest-cli==2.1.3
stack-data==0.6.3
stdiomask==0.0.6
stem==1.8.2
systemd-python==235
Tempita==0.5.2
tinycss2==1.2.1
tls-client==1.0.1
tornado==6.3.3
tqdm==4.65.0
traitlets==5.14.1
trio==0.22.2
trio-websocket==0.10.3
typing_extensions==4.12.2
urllib3==2.0.3
urwid-mitmproxy==2.1.2.1
wcwidth==0.2.6
webdriver-manager==4.0.2
webencodings==0.5.1
websocket-client==1.8.0
Werkzeug==2.3.7
wget==3.2
wikipedia==1.4.0
wsproto==1.2.0
xcffib==1.2.0
zope.interface==6.0
zstandard==0.22.0
